fix warning on max bans and change an iovec to an ircd_iovec
git-svn-id: http://svn.ratbox.org/svnroot/ircd-ratbox/trunk@23279 b93f080a-4cfa-0310-a19a-80a68f531ef9

+++ modules/core/m_mode.c
@@ -407,7 +407,7 @@ add_id(struct Client *source_p, struct Channel *chptr, const char *banid,
 	 */
 	if(MyClient(source_p))
 	{
-		if((ircd_dlink_list_length(&chptr->banlist) + ircd_dlink_list_length(&chptr->exceptlist) + ircd_dlink_list_length(&chptr->invexlist)) >= ConfigChannel.max_bans)
+		if((ircd_dlink_list_length(&chptr->banlist) + ircd_dlink_list_length(&chptr->exceptlist) + ircd_dlink_list_length(&chptr->invexlist)) >= (unsigned long)ConfigChannel.max_bans)
 		{
 			sendto_one(source_p, POP_QUEUE, form_str(ERR_BANLISTFULL),
 				   me.name, source_p->name, chptr->chname, realban);
